Follow/unfollow buttons
Added a follow/unfollow button on the userpage. If user is logged in,
but does not follow another user, follow button is shown. If user is
logged in, but does follow another user, unfollow button is shown. If
user is not logged in, follow button is shown, but clicking it causes a
login window to appear.