Large update
Hashtags are now detected in Kluch text. Users can click on them, and it
will take them to a feed of Kluchs with given hashtag. Rules for
hashtags are identical to Hashtags on twitter, facebook, instagram and
so on. Hashtags have to be preceded by spaces (or be at the beginning of
text), can contain digits, a-Z characters and underscores and can begin
with digits.