Notifications
Users are now able to receive notifications when they are @mentioned in
kluchs or when they are followed. "notification/poll" endpoint returns
number of unread notifications and "notification" endpoint returns the
notifications. On the front-side, notifications are polled with ajax and
displayed next to "notifications" button. Clicking on it leads to a page
containing kluchs in which you are mentioned and your new followers (if
any).

Also, changed all tests for controllers to use Mockito.verify to check
for interactions with services.