Changed how feeds work
FeedConstruction now takes next/previous as arguments. What this means
varies from feed to feed, but in MOST cases next is an id of an element
that is the oldest - e.g. if you are scrolling through your feed, next
is going to keep pointing at the last element (because you are watching
your feed from latest to oldest). Changed most front-end to work with
this, as well as implemented this in the back-end. Some database
modifications were inevitable. Also simplified notification system.
There are no longer any Notification objects, users simply have a field
called "notifications" which contains the number of unread
notifications. Last but not least, follower/followee relationship is now
represented as a separate table in the database, see Follow object.

What is left is to implement a feed of users who followed you and
connect it to front-end.